Add libcrypto and new crypto.h and crypto.c files to the build system. Add sql columns to support password hashing. Remove some changes that aren't related to this branch. Fix HPMHooking: now it extracted the new authentication system.

+++ src/common/crypto.c
@@ -13,8 +13,8 @@ bool crypto_random_bytes(unsigned char *buf, int num)
     return true;
 }
 
-bool crypto_pbkdf2_hmac_sha512(const char *pass, int passlen,
-        const unsigned char *salt, int saltlen, int iter, int outlen,
+bool crypto_pbkdf2_hmac_sha512(const char *pass, size_t passlen,
+        const unsigned char *salt, size_t saltlen, int iter, size_t outlen,
         unsigned char *out)
 {
     if(1 != PKCS5_PBKDF2_HMAC(pass, passlen, salt, saltlen, iter,
